requirements.txt
asgiref==3.8.1
Django==5.0.4
django-smtp-ssl==1.0
djangorestframework==3.15.1
Faker==24.14.0
pillow==10.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.4.4
